gadproject/requirements.txt
asgiref==3.8.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
defusedxml==0.8.0rc2
Django==5.0.6
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==3.13.0
django-extensions==3.2.3
django-filter==24.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-yasg==1.21.7
gunicorn==20.1.0
idna==3.7
inflection==0.5.1
install==1.3.5
mysqlclient==2.2.4
oauthlib==3.2.2
packaging==24.1
pillow==10.3.0
pycparser==2.22
PyJWT==2.8.0
python-decouple==3.8
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==70.1.0
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
